🔧 Update PrimateController's update method
This commit updates the `update` method in the `PrimateController` class. It adds a console log statement to check the type of `this.service`. Additionally, it modifies the logic for retrieving and updating records based on the availability of certain service methods. If either `get` or `update` methods are available in `this.service`, they will be used; otherwise, fallback methods from `PrimateService` will be used instead.