🔧 Add conditional check for options.resolveWhere
This commit adds a conditional check for `options.resolveWhere` in the `PrimateService` class. If `options.resolveWhere` is truthy, it will be used to set the `args.where` value; otherwise, the default `PrimateService.resolveWhere` method will be used. This change ensures flexibility in resolving the "where" parameter based on different options passed to the service.