🔧 Add user to options in PrimateController
The commit adds the user payload to the options object in the PrimateController class. This allows for easier access to the user information throughout the controller.

🐒 Update filterAllQuery in PrimateService

The commit updates the filterAllQuery function in the PrimateService class. It now checks if a filterAllQueryObject is returned and assigns it to queryObject if it exists.

🔍 Remove console.log statement

The commit removes a console.log statement from the fetchs loop in the PrimateService class.