✨ Add query parameters to the query object
This commit adds functionality to check if a query parameter is in the fields of the model and, if so, add it to the query. It iterates over the model fields and checks if the query has a field that matches. If there is a match, it adds the field to the query object.