🔧 Add joi dependency and implement validateSchema function
This commit adds the "joi" dependency to the package.json file. It also introduces a new function called "validateSchema" in the utils.js file. This function is used to validate data against a schema, which is dynamically imported based on the entity name. If no schema exists for the entity, validation will return true.