✨ Add GitHub Action for Retype
This commit adds a new GitHub Action workflow file `retype-build-action.yml` that sets up a workflow for Retype. The workflow is triggered on push events to the `retype` branch and manually through the workflow dispatch event. It runs on Ubuntu latest and has permissions to write contents. The steps include checking out the repository and using the latest version of the Retype build action.

#github #action #retype