🔄 Rename workflow file and update GitHub Action for Retype
The commit renames the workflow file from "retype-build-action.yml" to "retype-publish-page.yml". It also updates the name of the GitHub Action to "Publish Retype powered website to GitHub Pages". Additionally, it modifies the steps in the workflow by updating the version of `actions/checkout` and adding a new step using `retypeapp/action-github-pages` with an option to update the branch.