🔧 Update package version and dependencies
The package.json file was updated to increase the version number from 1.3.15-o1 to 1.3.16-o1. Additionally, the dependencies "@prisma/client" and "prisma" were updated from version 5.7.1 to 5.20.0.

📝 Modify setupRoute method parameter type

In the src/primate.js file, the parameter type for the router in the setupRoute method was changed from "express.Router" to just "Router".